Front move Java archunit in Spring boot section
