+++ src/main/webapp/app/springboot/domain/SpringBootService.ts
@@ -8,6 +8,7 @@ export interface SpringBootService {
   addSpringBootWebfluxNetty(project: Project): Promise<void>;
   addSpringBootActuator(project: Project): Promise<void>;
   addSpringDoc(project: Project): Promise<void>;
+  addJavaArchunit(project: Project): Promise<void>;
 
   addJWT(project: Project): Promise<void>;
   addBasicAuthJWT(project: Project): Promise<void>;
