+++ src/main/webapp/app/springboot/primary/generator/project-generator/ProjectGenerator.vue
@@ -47,12 +47,6 @@
       </div>
       <div>
         <GeneratorButtonVue :label="'Java Base'" :service="'java-base'" :selector-prefix="selectorPrefix" @click.prevent="addJavaBase" />
-        <GeneratorButtonVue
-          :label="'Java Archunit'"
-          :service="'java-archunit'"
-          :selector-prefix="selectorPrefix"
-          @click.prevent="addJavaArchunit"
-        />
       </div>
       <div>
         <GeneratorButtonVue
