+++ src/test/javascript/cypress/integration/Generator.spec.ts
@@ -28,7 +28,6 @@ describe('Generator', () => {
     cy.get(projectGeneratorSelector('add-sonar-java-backend-button')).contains('Sonar Backend');
     cy.get(projectGeneratorSelector('add-sonar-java-backend-and-frontend-button')).contains('Sonar Backend+Frontend');
     cy.get(projectGeneratorSelector('add-java-base-button')).contains('Java Base');
-    cy.get(projectGeneratorSelector('add-java-archunit-button')).contains('Java Archunit');
     cy.get(projectGeneratorSelector('add-frontend-maven-plugin-button')).contains('Frontend Maven Plugin');
     cy.get(projectGeneratorSelector('add-download-button')).should('not.exist');
   });
@@ -48,7 +47,7 @@ describe('Generator', () => {
     cy.get(springBootGeneratorSelector('add-spring-boot-webflux-netty-button')).contains('Spring Webflux Netty');
     cy.get(springBootGeneratorSelector('add-spring-boot-actuator-button')).contains('Spring Boot Actuator');
     cy.get(springBootGeneratorSelector('add-spring-doc-button')).contains('Spring Doc');
-
+    cy.get(springBootGeneratorSelector('add-java-archunit-button')).contains('Java Archunit');
     cy.get(springBootGeneratorSelector('add-aop-logging-button')).contains('AOP Logging');
     cy.get(springBootGeneratorSelector('add-logstash-button')).contains('Logstash');
