+++ src/test/javascript/spec/springboot/secondary/ProjectRepository.spec.ts
@@ -153,24 +153,6 @@ describe('ProjectRepository', () => {
     expect(projectFolder).toBe(PROJECT_FOLDER);
   });
 
-  it('should add JavaArchunit', async () => {
-    const projectHistoryService = stubProjectHistoryService();
-    const axiosHttpStub = stubAxiosHttp();
-    const projectStoreStub = stubProjectStore();
-    axiosHttpStub.post.resolves();
-    const projectRepository = new ProjectRepository(axiosHttpStub, projectHistoryService, projectStoreStub);
-    const project: Project = createProject({ folder: PROJECT_FOLDER });
-
-    await projectRepository.addJavaArchunit(project);
-
-    const expectedRestProject: RestProject = toRestProject(project);
-    const [uri, payload] = axiosHttpStub.post.getCall(0).args;
-    expect(uri).toBe('api/servers/java/arch');
-    expect(payload).toEqual<RestProject>(expectedRestProject);
-    const [projectFolder] = projectHistoryService.get.getCall(0).args;
-    expect(projectFolder).toBe(PROJECT_FOLDER);
-  });
-
   it('should add Frontend Maven Plugin', async () => {
     const projectHistoryService = stubProjectHistoryService();
     const axiosHttpStub = stubAxiosHttp();
